Trait automerge::op_observer::BranchableObserver

pub trait BranchableObserver {
    fn branch(&self) -> Self;
    fn merge(&mut self, other: &Self);
}

An observer which can be branched

This is used when observing operations in a transaction. In this case branch will be called at the beginning of the transaction to return a new observer and then merge will be called with the branched observer as other when the transaction is comitted.

Required Methods

fn branch(&self) -> Self

Branch of a new op_observer later to be merged

Called when creating a new transaction. Observer branch will be merged on commit() or thrown away on rollback()

fn merge(&mut self, other: &Self)

Merge observed information from a transaction.

Called by AutoCommit on commit()

other: Another Op Observer of the same type