Struct automerge::AutoCommitWithObs

pub struct AutoCommitWithObs<Obs: Observation> { /* private fields */ }

An automerge document that automatically manages transactions.

Implementations

impl<O: Observation> AutoCommitWithObs<O>

pub fn unobserved() -> AutoCommitWithObs<UnObserved>

impl AutoCommitWithObs<UnObserved>

pub fn new() -> AutoCommit

pub fn load(data: &[u8]) -> Result<Self, AutomergeError>

impl<Obs: OpObserver> AutoCommitWithObs<Observed<Obs>>

pub fn observer(&mut self) -> &mut Obs

impl<Obs: Observation + Clone> AutoCommitWithObs<Obs>

pub fn fork(&mut self) -> Self

pub fn fork_at(&mut self, heads: &[ChangeHash]) -> Result<Self, AutomergeError>

impl<Obs: Observation> AutoCommitWithObs<Obs>

pub fn with_observer<Obs2: OpObserver>(
self,
op_observer: Obs2
) -> AutoCommitWithObs<Observed<Obs2>>

pub fn with_actor(self, actor: ActorId) -> Self

pub fn set_actor(&mut self, actor: ActorId) -> &mut Self

pub fn get_actor(&self) -> &ActorId

pub fn with_encoding(self, encoding: TextEncoding) -> Self

pub fn load_incremental(&mut self, data: &[u8]) -> Result<usize, AutomergeError>

pub fn apply_changes(
&mut self,
changes: impl IntoIterator<Item = Change>
) -> Result<(), AutomergeError>

pub fn merge<Obs2: Observation>(
&mut self,
other: &mut AutoCommitWithObs<Obs2>
) -> Result<Vec<ChangeHash>, AutomergeError>

Takes all the changes in other which are not in self and applies them

pub fn save(&mut self) -> Vec<u8>

pub fn save_nocompress(&mut self) -> Vec<u8>

pub fn save_incremental(&mut self) -> Vec<u8>

pub fn get_missing_deps(&mut self, heads: &[ChangeHash]) -> Vec<ChangeHash>

pub fn get_last_local_change(&mut self) -> Option<&Change>

pub fn get_changes(
&mut self,
have_deps: &[ChangeHash]
) -> Result<Vec<&Change>, AutomergeError>

pub fn get_change_by_hash(&mut self, hash: &ChangeHash) -> Option<&Change>

pub fn get_changes_added<'a>(&mut self, other: &'a mut Self) -> Vec<&'a Change>

pub fn import(&self, s: &str) -> Result<(ExId, ObjType), AutomergeError>

pub fn dump(&mut self)

pub fn generate_sync_message(
&mut self,
sync_state: &mut State
) -> Option<Message>

pub fn receive_sync_message(
&mut self,
sync_state: &mut State,
message: Message
) -> Result<(), AutomergeError>

pub fn visualise_optree(&self, objects: Option<Vec<ExId>>) -> String

Return a graphviz representation of the opset.

Arguments

objects: An optional list of object IDs to display, if not specified all objects are visualised

pub fn get_heads(&mut self) -> Vec<ChangeHash>

Get the current heads of the document.

This closes the transaction first, if one is in progress.

pub fn commit(&mut self) -> Option<ChangeHash>

Commit any uncommitted changes

Returns None if there were no operations to commit

pub fn commit_with(&mut self, options: CommitOptions) -> Option<ChangeHash>

Commit the current operations with some options.

Returns None if there were no operations to commit

let mut doc = AutoCommit::new();
doc.put_object(&ROOT, "todos", ObjType::List).unwrap();
let now = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH).unwrap().as_secs() as
i64;
doc.commit_with(CommitOptions::default().with_message("Create todos list").with_time(now));

pub fn rollback(&mut self) -> usize

pub fn empty_change(&mut self, options: CommitOptions) -> ChangeHash

Generate an empty change

The main reason to do this is if you wish to create a “merge commit” which has all the current heads of the documents as dependencies but you have no new operations to create.

Because this structure is an “autocommit” there may actually be outstanding operations to submit. If this is the case this function will create two changes, one with the outstanding operations and a new one with no operations. The returned ChangeHash will always be the hash of the empty change.