Trait automerge::OpObserver

pub trait OpObserver: Default + Clone {
    fn insert(
        &mut self,
        parents: Parents<'_>,
        objid: ExId,
        index: usize,
        tagged_value: (Value<'_>, ExId)
    );
    fn put(
        &mut self,
        parents: Parents<'_>,
        objid: ExId,
        prop: Prop,
        tagged_value: (Value<'_>, ExId),
        conflict: bool
    );
    fn increment(
        &mut self,
        parents: Parents<'_>,
        objid: ExId,
        prop: Prop,
        tagged_value: (i64, ExId)
    );
    fn delete(&mut self, parents: Parents<'_>, objid: ExId, prop: Prop);
    fn merge(&mut self, other: &Self);

    fn branch(&self) -> Self { ... }
}

An observer of operations applied to the document.

Required Methods

fn insert(
&mut self,
parents: Parents<'_>,
objid: ExId,
index: usize,
tagged_value: (Value<'_>, ExId)
)

A new value has been inserted into the given object.

parents: A parents iterator that can be used to collect path information

objid: the object that has been inserted into.

index: the index the new value has been inserted at.

tagged_value: the value that has been inserted and the id of the operation that did the insert.

fn put(
&mut self,
parents: Parents<'_>,
objid: ExId,
prop: Prop,
tagged_value: (Value<'_>, ExId),
conflict: bool
)

A new value has been put into the given object.

parents: A parents iterator that can be used to collect path information

objid: the object that has been put into.

prop: the prop that the value as been put at.

tagged_value: the value that has been put into the object and the id of the operation that did the put.

conflict: whether this put conflicts with other operations.

fn increment(
&mut self,
parents: Parents<'_>,
objid: ExId,
prop: Prop,
tagged_value: (i64, ExId)
)

A counter has been incremented.

parents: A parents iterator that can be used to collect path information

objid: the object that contains the counter.

prop: they prop that the chounter is at.

tagged_value: the amount the counter has been incremented by, and the the id of the increment operation.

fn delete(&mut self, parents: Parents<'_>, objid: ExId, prop: Prop)

A value has beeen deleted.

parents: A parents iterator that can be used to collect path information

objid: the object that has been deleted in.

prop: the prop of the value that has been deleted.

fn merge(&mut self, other: &Self)

Merge observed information from a transaction.

Called by AutoCommit on commit()

other: Another Op Observer of the same type

Provided Methods

fn branch(&self) -> Self

Branch of a new op_observer later to be merged

Called by AutoCommit when creating a new transaction. Observer branch will be merged on commit() or thrown away on rollback()