Crate automerge

Modules

sync

transaction

Structs

ActorId

An actor id is a sequence of bytes. By default we use a uuid which can be nicely stack allocated.

AutoCommitWithObs

An automerge document that automatically manages transactions.

AutoSerde

A wrapper type which implements serde::Serialize for an Automerge.

Automerge

An automerge document.

Change

ChangeHash

The sha256 hash of a change.

ExpandedChange

A change represents a group of operations performed by an actor.

InvalidActorId

InvalidChangeHashSlice

Keys

KeysAt

ListRange

ListRangeAt

MapRange

MapRangeAt

Parents

Values

VecOpObserver

Capture operations into a Vec and store them as patches.

Enums

AutomergeError

LoadChangeError

ObjId

ObjType

OpType

Patch

A notification to the application that something has changed in a document.

Prop

ScalarValue

Value

Constants

ROOT

Traits

OpObserver

An observer of operations applied to the document.

Type Definitions

AutoCommit