Crate automerge

Modules

sync

transaction

Structs

ActorId

An actor id is a sequence of bytes. By default we use a uuid which can be nicely stack allocated.

ApplyOptions

AutoCommit

An automerge document that automatically manages transactions.

Automerge

An automerge document.

Change

A change represents a group of operations performed by an actor.

ChangeHash

The sha256 hash of a change.

ExpandedChange

A change represents a group of operations performed by an actor.

Keys

KeysAt

ListRange

ListRangeAt

MapRange

MapRangeAt

Parents

Values

VecOpObserver

Capture operations into a Vec and store them as patches.

Enums

AutomergeError

DecodingError

The error type for decoding operations.

EncodingError

The error type for encoding operations.

InvalidChangeError

ObjId

ObjType

OpType

Patch

A notification to the application that something has changed in a document.

Prop

ScalarValue

Value

Constants

ROOT

Traits

OpObserver

An observer of operations applied to the document.