Crate automerge

Modules

value_ref

Structs

Automerge

The core automerge document.

AutomergeBuilder

A more advanced way of building an Automerge from constituent parts.

Backend

Change

ChangeHash

Frontend

LocalChange

Path

Enums

AutomergeError

BackendError

FrontendError

InvalidChangeRequest

InvalidPatch

MapType

ObjType

Primitive

A primitive value, leaf nodes of the document tree.

ScalarValue

SequenceType

Value

A composite value, composing maps, tables, lists, text and primitives.

Traits

MutableDocument