Trait automerge::op_observer::OpObserver

pub trait OpObserver {
    fn insert<R: ReadDoc>(
        &mut self,
        doc: &R,
        objid: ExId,
        index: usize,
        tagged_value: (Value<'_>, ExId)
    );
    fn splice_text<R: ReadDoc>(
        &mut self,
        _doc: &R,
        _objid: ExId,
        _index: usize,
        _value: &str
    );
    fn put<R: ReadDoc>(
        &mut self,
        doc: &R,
        objid: ExId,
        prop: Prop,
        tagged_value: (Value<'_>, ExId),
        conflict: bool
    );
    fn expose<R: ReadDoc>(
        &mut self,
        doc: &R,
        objid: ExId,
        prop: Prop,
        tagged_value: (Value<'_>, ExId),
        conflict: bool
    );
    fn increment<R: ReadDoc>(
        &mut self,
        doc: &R,
        objid: ExId,
        prop: Prop,
        tagged_value: (i64, ExId)
    );
    fn delete_map<R: ReadDoc>(&mut self, doc: &R, objid: ExId, key: &str);
    fn delete_seq<R: ReadDoc>(
        &mut self,
        doc: &R,
        objid: ExId,
        index: usize,
        num: usize
    );

    fn flag_conflict<R: ReadDoc>(&mut self, _doc: &R, _objid: ExId, _prop: Prop) { ... }
    fn delete<R: ReadDoc>(&mut self, doc: &R, objid: ExId, prop: Prop) { ... }
    fn text_as_seq(&self) -> bool { ... }
}

An observer of operations applied to the document.

Required Methods

fn insert<R: ReadDoc>(
&mut self,
doc: &R,
objid: ExId,
index: usize,
tagged_value: (Value<'_>, ExId)
)

A new value has been inserted into the given object.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been inserted into.

index: the index the new value has been inserted at.

tagged_value: the value that has been inserted and the id of the operation that did the insert.

fn splice_text<R: ReadDoc>(
&mut self,
_doc: &R,
_objid: ExId,
_index: usize,
_value: &str
)

Some text has been spliced into a text object

fn put<R: ReadDoc>(
&mut self,
doc: &R,
objid: ExId,
prop: Prop,
tagged_value: (Value<'_>, ExId),
conflict: bool
)

A new value has been put into the given object.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been put into.

prop: the prop that the value as been put at.

tagged_value: the value that has been put into the object and the id of the operation that did the put.

conflict: whether this put conflicts with other operations.

fn expose<R: ReadDoc>(
&mut self,
doc: &R,
objid: ExId,
prop: Prop,
tagged_value: (Value<'_>, ExId),
conflict: bool
)

When a delete op exposes a previously conflicted value Similar to a put op - except for maps, lists and text, edits may already exist and need to be queried

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been put into.

prop: the prop that the value as been put at.

tagged_value: the value that has been put into the object and the id of the operation that did the put.

conflict: whether this put conflicts with other operations.

fn increment<R: ReadDoc>(
&mut self,
doc: &R,
objid: ExId,
prop: Prop,
tagged_value: (i64, ExId)
)

A counter has been incremented.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that contains the counter.

prop: they prop that the chounter is at.

tagged_value: the amount the counter has been incremented by, and the the id of the increment operation.

fn delete_map<R: ReadDoc>(&mut self, doc: &R, objid: ExId, key: &str)

A map value has beeen deleted.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been deleted in.

key: the map key to be deleted

fn delete_seq<R: ReadDoc>(&mut self, doc: &R, objid: ExId, index: usize, num: usize)

A one or more list values have beeen deleted.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been deleted in.

index: the index of the deletion

num: the number of sequential elements deleted

Provided Methods

fn flag_conflict<R: ReadDoc>(&mut self, _doc: &R, _objid: ExId, _prop: Prop)

Flag a new conflict on a value without changing it

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been put into.

prop: the prop that the value as been put at.

fn delete<R: ReadDoc>(&mut self, doc: &R, objid: ExId, prop: Prop)

A map value has beeen deleted.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been deleted in.

prop: the prop to be deleted

fn text_as_seq(&self) -> bool

Whether to call sequence methods or splice_text when encountering changes in text

Returns false by default