Struct automerge::AutoCommit

pub struct AutoCommit { /* private fields */ }

An automerge document that automatically manages transactions.

Implementations

impl AutoCommit

pub fn new() -> Self

pub fn with_actor(self, actor: ActorId) -> Self

pub fn set_actor(&mut self, actor: ActorId) -> &mut Self

pub fn get_actor(&self) -> &ActorId

pub fn fork(&mut self) -> Self

pub fn fork_at(&mut self, heads: &[ChangeHash]) -> Result<Self, AutomergeError>

pub fn load(data: &[u8]) -> Result<Self, AutomergeError>

pub fn load_with<Obs: OpObserver>(
data: &[u8],
options: ApplyOptions<'_, Obs>
) -> Result<Self, AutomergeError>

pub fn load_incremental(&mut self, data: &[u8]) -> Result<usize, AutomergeError>

pub fn load_incremental_with<'a, Obs: OpObserver>(
&mut self,
data: &[u8],
options: ApplyOptions<'a, Obs>
) -> Result<usize, AutomergeError>

pub fn apply_changes(
&mut self,
changes: Vec<Change>
) -> Result<(), AutomergeError>

pub fn apply_changes_with<Obs: OpObserver>(
&mut self,
changes: Vec<Change>,
options: ApplyOptions<'_, Obs>
) -> Result<(), AutomergeError>

pub fn merge(
&mut self,
other: &mut Self
) -> Result<Vec<ChangeHash>, AutomergeError>

Takes all the changes in other which are not in self and applies them

pub fn merge_with<'a, Obs: OpObserver>(
&mut self,
other: &mut Self,
options: ApplyOptions<'a, Obs>
) -> Result<Vec<ChangeHash>, AutomergeError>

Takes all the changes in other which are not in self and applies them

pub fn save(&mut self) -> Vec<u8>

pub fn save_incremental(&mut self) -> Vec<u8>

pub fn get_missing_deps(&mut self, heads: &[ChangeHash]) -> Vec<ChangeHash>

pub fn get_last_local_change(&mut self) -> Option<&Change>

pub fn get_changes(&mut self, have_deps: &[ChangeHash]) -> Vec<&Change>

pub fn get_change_by_hash(&mut self, hash: &ChangeHash) -> Option<&Change>

pub fn get_changes_added<'a>(&mut self, other: &'a mut Self) -> Vec<&'a Change>

pub fn import(&self, s: &str) -> Result<ExId, AutomergeError>

pub fn dump(&mut self)

pub fn generate_sync_message(
&mut self,
sync_state: &mut State
) -> Option<Message>

pub fn receive_sync_message(
&mut self,
sync_state: &mut State,
message: Message
) -> Result<(), AutomergeError>

pub fn receive_sync_message_with<'a, Obs: OpObserver>(
&mut self,
sync_state: &mut State,
message: Message,
options: ApplyOptions<'a, Obs>
) -> Result<(), AutomergeError>

pub fn visualise_optree(&self) -> String

pub fn get_heads(&mut self) -> Vec<ChangeHash>

Get the current heads of the document.

This closes the transaction first, if one is in progress.

pub fn commit(&mut self) -> ChangeHash

pub fn commit_with<Obs: OpObserver>(
&mut self,
options: CommitOptions<'_, Obs>
) -> ChangeHash

Commit the current operations with some options.

let mut doc = AutoCommit::new();
doc.put_object(&ROOT, "todos", ObjType::List).unwrap();
let now = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH).unwrap().as_secs() as
i64;
doc.commit_with::<()>(CommitOptions::default().with_message("Create todos list").with_time(now));

pub fn rollback(&mut self) -> usize