Struct automerge::AutoCommit

pub struct AutoCommit { /* private fields */ }

An automerge document that automatically manages transactions.

Creating, loading, merging and forking documents

A new document can be created with Self::new, which will create a document with a random ActorId. Existing documents can be loaded with Self::load.

If you have two documents and you want to merge the changes from one into the other you can use Self::merge.

If you have a document you want to split into two concurrent threads of execution you can use Self::fork. If you want to split a document from ealier in its history you can use Self::fork_at.

Reading values

Self implements ReadDoc, which provides methods for reading values from the document.

Modifying a document

This type implements Transactable directly, so you can modify it using methods from Transactable.

Synchronization

To synchronise call Self::sync which returns an implementation of SyncDoc

Patches, maintaining materialized views

AutoCommit allows you to generate Patches representing changes to the current state of the document which you can use to maintain a materialized view of the current state. There are several ways to use this. See the documentation on Self::diff for more details, but the key point to remember is that AutoCommit manages an internal “diff cursor” for you. This is a representation of the heads of the document last time you called diff_incremental but you can also manage it directly using Self::update_diff_cursor and Self::reset_diff_cursor.

Implementations

impl AutoCommit

pub fn new() -> AutoCommit

pub fn load(data: &[u8]) -> Result<Self, AutomergeError>

pub fn load_unverified_heads(data: &[u8]) -> Result<Self, AutomergeError>

pub fn load_with( data: &[u8], on_error: OnPartialLoad, mode: VerificationMode ) -> Result<Self, AutomergeError>

pub fn reset_diff_cursor(&mut self)

Erases the diff cursor created by Self::update_diff_cursor and no longer indexes changes to the document.

pub fn update_diff_cursor(&mut self)

Sets the Self::diff_cursor to current heads of the document and will begin building an index with every change moving forward.

If Self::diff is called with Self::diff_cursor as before and Self::get_heads as after - the index will be used

If the cursor is no longer needed it can be reset with Self::reset_diff_cursor

pub fn diff_cursor(&self) -> Vec<ChangeHash>

Returns the cursor set by Self::update_diff_cursor

pub fn make_patches(&self, patch_log: &mut PatchLog) -> Vec<Patch>

Generate the patches recorded in patch_log

pub fn diff( &mut self, before: &[ChangeHash], after: &[ChangeHash] ) -> Vec<Patch>

Generates a diff from before to after

By default the diff requires a sequental scan of all the ops in the doc.

To do a fast indexed diff before must equal Self::diff_cursor and after must equal Self::get_heads. The diff cursor is managed with Self::update_diff_cursor and Self::reset_diff_cursor

Managing the diff index has a small but non-zero overhead. It should be disabled if no longer needed. If a signifigantly large change is applied to the document it may be faster to reset the index before applying it, doing an unindxed diff afterwards and then reenable the index.

Arguments

before - heads from Self::get_heads() at beginning point in the documents history

after - heads from Self::get_heads() at ending point in the documents history.

Note: before and after do not have to be chronological. Document state can move backward. Normal use might look like:

Example

use automerge::{ AutoCommit };

let mut doc = AutoCommit::new(); // or AutoCommit::load(data)
// make some changes - use and update the index
let heads = doc.get_heads();
let diff_cursor = doc.diff_cursor();
let patches = doc.diff(&diff_cursor, &heads);
doc.update_diff_cursor();

See Self::diff_incremental for encapsulating this pattern.

pub fn diff_incremental(&mut self) -> Vec<Patch>

This is a convience function that encapsulates the following common pattern

use automerge::AutoCommit;
let mut doc = AutoCommit::new();
// make some changes
let heads = doc.get_heads();
let diff_cursor = doc.diff_cursor();
let patches = doc.diff(&diff_cursor, &heads);
doc.update_diff_cursor();

pub fn fork(&mut self) -> Self

pub fn fork_at(&mut self, heads: &[ChangeHash]) -> Result<Self, AutomergeError>

pub fn with_actor(self, actor: ActorId) -> Self

pub fn set_actor(&mut self, actor: ActorId) -> &mut Self

pub fn get_actor(&self) -> &ActorId

pub fn isolate(&mut self, heads: &[ChangeHash])

pub fn integrate(&mut self)

pub fn load_incremental(&mut self, data: &[u8]) -> Result<usize, AutomergeError>

Load an incremental save of a document.

Unlike load this imports changes into an existing document. It will work with both the output of Self::save and Self::save_incremental

The return value is the number of ops which were applied, this is not useful and will change in future.

pub fn apply_changes( &mut self, changes: impl IntoIterator<Item = Change> ) -> Result<(), AutomergeError>

pub fn merge( &mut self, other: &mut AutoCommit ) -> Result<Vec<ChangeHash>, AutomergeError>

Takes all the changes in other which are not in self and applies them

pub fn save(&mut self) -> Vec<u8>

Save the entirety of this document in a compact form.

pub fn save_with_options(&mut self, options: SaveOptions) -> Vec<u8>

pub fn save_and_verify(&mut self) -> Result<Vec<u8>, AutomergeError>

Save the document and attempt to load it before returning - slow!

pub fn save_nocompress(&mut self) -> Vec<u8>

Save this document, but don’t run it through DEFLATE afterwards

pub fn save_incremental(&mut self) -> Vec<u8>

Save the changes since the last call to Self::save`

The output of this will not be a compressed document format, but a series of individual changes. This is useful if you know you have only made a small change since the last save and you want to immediately send it somewhere (e.g. you’ve inserted a single character in a text object).

pub fn save_after(&mut self, heads: &[ChangeHash]) -> Vec<u8>

Save everything which is not a (transitive) dependency of heads

pub fn get_missing_deps(&mut self, heads: &[ChangeHash]) -> Vec<ChangeHash>

pub fn get_last_local_change(&mut self) -> Option<&Change>

Get the last change made by this documents actor ID

pub fn get_changes(&mut self, have_deps: &[ChangeHash]) -> Vec<&Change>

pub fn get_change_by_hash(&mut self, hash: &ChangeHash) -> Option<&Change>

pub fn get_changes_added<'a>(&mut self, other: &'a mut Self) -> Vec<&'a Change>

Get changes in other that are not in `self

pub fn visualise_optree(&self, objects: Option<Vec<ExId>>) -> String

Return a graphviz representation of the opset.

Arguments

objects: An optional list of object IDs to display, if not specified all objects are visualised

pub fn get_heads(&mut self) -> Vec<ChangeHash>

Get the current heads of the document.

This closes the transaction first, if one is in progress.

pub fn set_text_rep(&mut self, text_rep: TextRepresentation)

pub fn get_text_rep(&mut self) -> TextRepresentation

pub fn with_text_rep(self, text_rep: TextRepresentation) -> Self

pub fn commit(&mut self) -> Option<ChangeHash>

Commit any uncommitted changes

Returns None if there were no operations to commit

pub fn commit_with(&mut self, options: CommitOptions) -> Option<ChangeHash>

Commit the current operations with some options.

Returns None if there were no operations to commit

let mut doc = AutoCommit::new();
doc.put_object(&ROOT, "todos", ObjType::List).unwrap();
let now = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH).unwrap().as_secs() as
i64;
doc.commit_with(CommitOptions::default().with_message("Create todos list").with_time(now));

pub fn rollback(&mut self) -> usize

Remove any changes that have been made in the current transaction from the document

pub fn empty_change(&mut self, options: CommitOptions) -> ChangeHash

Generate an empty change

The main reason to do this is if you wish to create a “merge commit” which has all the current heads of the documents as dependencies but you have no new operations to create.

Because this structure is an “autocommit” there may actually be outstanding operations to submit. If this is the case this function will create two changes, one with the outstanding operations and a new one with no operations. The returned ChangeHash will always be the hash of the empty change.

pub fn sync(&mut self) -> impl SyncDoc + '_

An implementation of crate::sync::SyncDoc for this autocommit

This ensures that any outstanding transactions for this document are committed before taking part in the sync protocol

pub fn hash_for_opid(&self, opid: &ExId) -> Option<ChangeHash>

Get the hash of the change that contains the given opid.

Returns none if the opid:

is the root object id

does not exist in this document

is for an operation in a transaction

pub fn hydrate(&self, heads: Option<&[ChangeHash]>) -> Value