Struct automerge::Automerge

pub struct Automerge { /* private fields */ }

An automerge document.

Implementations

impl Automerge

pub fn new() -> Self

Create a new document with a random actor id.

pub fn with_actor(self, actor: ActorId) -> Self

Set the actor id for this document.

pub fn set_actor(&mut self, actor: ActorId) -> &mut Self

Set the actor id for this document.

pub fn get_actor(&self) -> &ActorId

Get the current actor id of this document.

pub fn transaction(&mut self) -> Transaction<'_>

Start a transaction.

pub fn transact<F, O, E>(&mut self, f: F) -> Result<O, E> where
F: FnOnce(&mut Transaction<'_>) -> Result<O, E>,

Run a transaction on this document in a closure, automatically handling commit or rollback afterwards.

pub fn transact_with<'a, F, O, E, C, Obs>(&mut self, c: C, f: F) -> Result<O, E> where
F: FnOnce(&mut Transaction<'_>) -> Result<O, E>,
C: FnOnce(&O) -> CommitOptions<'a, Obs>,
Obs: 'a + OpObserver,

Like Self::transact but with a function for generating the commit options.

pub fn fork(&self) -> Self

Fork this document at the current point for use by a different actor.

pub fn fork_at(&self, heads: &[ChangeHash]) -> Result<Self, AutomergeError>

Fork this document at the give heads

pub fn parent_object<O: AsRef<ExId>>(&self, obj: O) -> Option<(ExId, Prop)>

Get the object id of the object that contains this object and the prop that this object is at in that object.

pub fn parents(&self, obj: ExId) -> Parents<'_>

Get an iterator over the parents of an object.

pub fn path_to_object<O: AsRef<ExId>>(&self, obj: O) -> Vec<(ExId, Prop)>

pub fn keys<O: AsRef<ExId>>(&self, obj: O) -> Keys<'_, '_>

Get the keys of the object obj.

For a map this returns the keys of the map. For a list this returns the element ids (opids) encoded as strings.

pub fn keys_at<O: AsRef<ExId>>(
&self,
obj: O,
heads: &[ChangeHash]
) -> KeysAt<'_, '_>

Historical version of keys.

pub fn map_range<O: AsRef<ExId>, R: RangeBounds<String>>(
&self,
obj: O,
range: R
) -> MapRange<'_, R>

Iterate over the keys and values of the map obj in the given range.

pub fn map_range_at<O: AsRef<ExId>, R: RangeBounds<String>>(
&self,
obj: O,
range: R,
heads: &[ChangeHash]
) -> MapRangeAt<'_, R>

Historical version of map_range.

pub fn list_range<O: AsRef<ExId>, R: RangeBounds<usize>>(
&self,
obj: O,
range: R
) -> ListRange<'_, R>

Iterate over the indexes and values of the list obj in the given range.

pub fn list_range_at<O: AsRef<ExId>, R: RangeBounds<usize>>(
&self,
obj: O,
range: R,
heads: &[ChangeHash]
) -> ListRangeAt<'_, R>

Historical version of list_range.

pub fn values<O: AsRef<ExId>>(&self, obj: O) -> Values<'_>

pub fn values_at<O: AsRef<ExId>>(
&self,
obj: O,
heads: &[ChangeHash]
) -> Values<'_>

pub fn length<O: AsRef<ExId>>(&self, obj: O) -> usize

Get the length of the given object.

pub fn length_at<O: AsRef<ExId>>(&self, obj: O, heads: &[ChangeHash]) -> usize

Historical version of length.

pub fn object_type<O: AsRef<ExId>>(&self, obj: O) -> Option<ObjType>

Get the type of this object, if it is an object.

pub fn text<O: AsRef<ExId>>(&self, obj: O) -> Result<String, AutomergeError>

Get the string represented by the given text object.

pub fn text_at<O: AsRef<ExId>>(
&self,
obj: O,
heads: &[ChangeHash]
) -> Result<String, AutomergeError>

Historical version of text.

pub fn get<O: AsRef<ExId>, P: Into<Prop>>(
&self,
obj: O,
prop: P
) -> Result<Option<(Value<'_>, ExId)>, AutomergeError>

Get a value out of the document.

Returns both the value and the id of the operation that created it, useful for handling conflicts and serves as the object id if the value is an object.

pub fn get_at<O: AsRef<ExId>, P: Into<Prop>>(
&self,
obj: O,
prop: P,
heads: &[ChangeHash]
) -> Result<Option<(Value<'_>, ExId)>, AutomergeError>

Historical version of get.

pub fn get_all<O: AsRef<ExId>, P: Into<Prop>>(
&self,
obj: O,
prop: P
) -> Result<Vec<(Value<'_>, ExId)>, AutomergeError>

Get all conflicting values out of the document at this prop that conflict.

Returns both the value and the id of the operation that created it, useful for handling conflicts and serves as the object id if the value is an object.

pub fn get_all_at<O: AsRef<ExId>, P: Into<Prop>>(
&self,
obj: O,
prop: P,
heads: &[ChangeHash]
) -> Result<Vec<(Value<'_>, ExId)>, AutomergeError>

Historical version of get_all.

pub fn load(data: &[u8]) -> Result<Self, AutomergeError>

Load a document.

pub fn load_with<Obs: OpObserver>(
data: &[u8],
options: ApplyOptions<'_, Obs>
) -> Result<Self, AutomergeError>

Load a document.

pub fn load_incremental(&mut self, data: &[u8]) -> Result<usize, AutomergeError>

Load an incremental save of a document.

pub fn load_incremental_with<Obs: OpObserver>(
&mut self,
data: &[u8],
options: ApplyOptions<'_, Obs>
) -> Result<usize, AutomergeError>

Load an incremental save of a document.

pub fn apply_changes(
&mut self,
changes: impl IntoIterator<Item = Change>
) -> Result<(), AutomergeError>

Apply changes to this document.

pub fn apply_changes_with<I: IntoIterator<Item = Change>, Obs: OpObserver>(
&mut self,
changes: I,
options: ApplyOptions<'_, Obs>
) -> Result<(), AutomergeError>

Apply changes to this document.

pub fn merge(
&mut self,
other: &mut Self
) -> Result<Vec<ChangeHash>, AutomergeError>

Takes all the changes in other which are not in self and applies them

pub fn merge_with<'a, Obs: OpObserver>(
&mut self,
other: &mut Self,
options: ApplyOptions<'a, Obs>
) -> Result<Vec<ChangeHash>, AutomergeError>

Takes all the changes in other which are not in self and applies them

pub fn save(&mut self) -> Vec<u8>

Save the entirety of this document in a compact form.

pub fn save_incremental(&mut self) -> Vec<u8>

Save the changes since last save in a compact form.

pub fn get_missing_deps(&self, heads: &[ChangeHash]) -> Vec<ChangeHash>

Get the hashes of the changes in this document that aren’t transitive dependencies of the given heads.

pub fn get_changes(
&self,
have_deps: &[ChangeHash]
) -> Result<Vec<&Change>, AutomergeError>

pub fn get_last_local_change(&self) -> Option<&Change>

Get the last change this actor made to the document.

pub fn get_change_by_hash(&self, hash: &ChangeHash) -> Option<&Change>

Get a change by its hash.

pub fn get_changes_added<'a>(&self, other: &'a Self) -> Vec<&'a Change>

Get the changes that the other document added compared to this document.

pub fn get_heads(&self) -> Vec<ChangeHash>

Get the heads of this document.

pub fn import(&self, s: &str) -> Result<ExId, AutomergeError>

pub fn dump(&self)

pub fn visualise_optree(&self) -> String

impl Automerge

pub fn generate_sync_message(&self, sync_state: &mut State) -> Option<Message>

pub fn receive_sync_message(
&mut self,
sync_state: &mut State,
message: Message
) -> Result<(), AutomergeError>

pub fn receive_sync_message_with<'a, Obs: OpObserver>(
&mut self,
sync_state: &mut State,
message: Message,
options: ApplyOptions<'a, Obs>
) -> Result<(), AutomergeError>