Struct automerge::Automerge

pub struct Automerge { /* private fields */ }

An automerge document which does not manage transactions for you.

Creating, loading, merging and forking documents

A new document can be created with Self::new, which will create a document with a random ActorId. Existing documents can be loaded with Self::load, or Self::load_with.

If you have two documents and you want to merge the changes from one into the other you can use Self::merge or Self::merge_and_log_patches.

If you have a document you want to split into two concurrent threads of execution you can use Self::fork. If you want to split a document from ealier in its history you can use Self::fork_at.

Reading values

Self implements ReadDoc, which provides methods for reading values from the document.

Modifying a document (Transactions)

Automerge provides an interface for viewing and modifying automerge documents which does not manage transactions for you. To create changes you use either Automerge::transaction or Automerge::transact (or the _with variants).

Sync

This type implements crate::sync::SyncDoc

Implementations

impl Automerge

pub fn new() -> Self

Create a new document with a random actor id.

pub fn is_empty(&self) -> bool

Whether this document has any operations

pub fn with_actor(self, actor: ActorId) -> Self

Set the actor id for this document.

pub fn set_actor(&mut self, actor: ActorId) -> &mut Self

Set the actor id for this document.

pub fn get_actor(&self) -> &ActorId

Get the current actor id of this document.

pub fn transaction(&mut self) -> Transaction<'_>

Start a transaction.

pub fn transaction_log_patches( &mut self, patch_log: PatchLog ) -> Transaction<'_>

Start a transaction which records changes in a PatchLog

pub fn transaction_at( &mut self, patch_log: PatchLog, heads: &[ChangeHash] ) -> Transaction<'_>

Start a transaction isolated at a given heads

pub fn transact<F, O, E>(&mut self, f: F) -> Result<O, E>where F: FnOnce(&mut Transaction<'_>) -> Result<O, E>,

Run a transaction on this document in a closure, automatically handling commit or rollback afterwards.

pub fn transact_with<F, O, E, C>(&mut self, c: C, f: F) -> Result<O, E>where F: FnOnce(&mut Transaction<'_>) -> Result<O, E>, C: FnOnce(&O) -> CommitOptions,

Like Self::transact but with a function for generating the commit options.

pub fn transact_and_log_patches<F, O, E>( &mut self, text_rep: TextRepresentation, f: F ) -> Result<O, E>where F: FnOnce(&mut Transaction<'_>) -> Result<O, E>,

Run a transaction on this document in a closure, collecting patches, automatically handling commit or rollback afterwards.

The collected patches are available in the return value of Transaction::commit

pub fn transact_and_log_patches_with<F, O, E, C>( &mut self, text_rep: TextRepresentation, c: C, f: F ) -> Result<O, E>where F: FnOnce(&mut Transaction<'_>) -> Result<O, E>, C: FnOnce(&O) -> CommitOptions,

Like Self::transact_and_log_patches but with a function for generating the commit options

pub fn empty_commit(&mut self, opts: CommitOptions) -> ChangeHash

Generate an empty change

The main reason to do this is if you want to create a “merge commit”, which is a change that has all the current heads of the document as dependencies.

pub fn fork(&self) -> Self

Fork this document at the current point for use by a different actor.

This will create a new actor ID for the forked document

pub fn fork_at(&self, heads: &[ChangeHash]) -> Result<Self, AutomergeError>

Fork this document at the given heads

This will create a new actor ID for the forked document

pub fn load(data: &[u8]) -> Result<Self, AutomergeError>

Load a document.

pub fn load_unverified_heads(data: &[u8]) -> Result<Self, AutomergeError>

Load a document without verifying the head hashes

This is useful for debugging as it allows you to examine a corrupted document.

pub fn load_with( data: &[u8], on_error: OnPartialLoad, mode: VerificationMode, patch_log: &mut PatchLog ) -> Result<Self, AutomergeError>

Load a document, with options

Arguments

data - The data to load

on_error - What to do if the document is only partially loaded. This can happen if some prefix of data contains valid data.

mode - Whether to verify the head hashes after loading

patch_log - A PatchLog to log the changes required to materialize the current state of the document once loaded

pub fn make_patches(&self, patch_log: &mut PatchLog) -> Vec<Patch>

Create the patches from a PatchLog

See the documentation for PatchLog for more details on this

pub fn current_state(&self, text_rep: TextRepresentation) -> Vec<Patch>

Get a set of Patches which materialize the current state of the document

This is a convienence method for doc.diff(&[], current_heads)

pub fn load_incremental(&mut self, data: &[u8]) -> Result<usize, AutomergeError>

Load an incremental save of a document.

Unlike load this imports changes into an existing document. It will work with both the output of Self::save and Self::save_after

The return value is the number of ops which were applied, this is not useful and will change in future.

pub fn load_incremental_log_patches( &mut self, data: &[u8], patch_log: &mut PatchLog ) -> Result<usize, AutomergeError>

Like Self::load_incremental but log the changes to the current state of the document to PatchLog

pub fn apply_changes( &mut self, changes: impl IntoIterator<Item = Change> ) -> Result<(), AutomergeError>

Apply changes to this document.

This is idemptotent in the sense that if a change has already been applied it will be ignored.

pub fn apply_changes_log_patches<I: IntoIterator<Item = Change>>( &mut self, changes: I, patch_log: &mut PatchLog ) -> Result<(), AutomergeError>

Like Self::apply_changes but log the resulting changes to the current state of the document to patch_log

pub fn merge( &mut self, other: &mut Self ) -> Result<Vec<ChangeHash>, AutomergeError>

Takes all the changes in other which are not in self and applies them

pub fn merge_and_log_patches( &mut self, other: &mut Self, patch_log: &mut PatchLog ) -> Result<Vec<ChangeHash>, AutomergeError>

Takes all the changes in other which are not in self and applies them whilst logging the resulting changes to the current state of the document to patch_log

pub fn save_with_options(&self, options: SaveOptions) -> Vec<u8>

Save the entirety of this document in a compact form.

pub fn save(&self) -> Vec<u8>

Save the entirety of this document in a compact form.

pub fn save_and_verify(&self) -> Result<Vec<u8>, AutomergeError>

Save the document and attempt to load it before returning - slow!

pub fn save_nocompress(&self) -> Vec<u8>

Save this document, but don’t run it through DEFLATE afterwards

pub fn save_after(&self, heads: &[ChangeHash]) -> Vec<u8>

Save the changes since the given heads

The output of this will not be a compressed document format, but a series of individual changes. This is useful if you know you have only made a small change since the last save and you want to immediately send it somewhere (e.g. you’ve inserted a single character in a text object).

pub fn get_last_local_change(&self) -> Option<&Change>

Get the last change this actor made to the document.

pub fn dump(&self)

pub fn visualise_optree(&self, objects: Option<Vec<ExId>>) -> String

Return a graphviz representation of the opset.

Arguments

objects: An optional list of object IDs to display, if not specified all objects are visualised

pub fn diff( &self, before_heads: &[ChangeHash], after_heads: &[ChangeHash], text_rep: TextRepresentation ) -> Vec<Patch>

Create patches representing the change in the current state of the document between the ‘before’ and ‘after’ heads. If the arguments are reverse it will observe the same changes in the opposite order.

pub fn get_heads(&self) -> Vec<ChangeHash>

Get the heads of this document.

pub fn get_changes(&self, have_deps: &[ChangeHash]) -> Vec<&Change>

pub fn get_changes_added<'a>(&self, other: &'a Self) -> Vec<&'a Change>

Get changes in other that are not in `self

pub fn hash_for_opid(&self, exid: &ExId) -> Option<ChangeHash>

Get the hash of the change that contains the given opid.

Returns none if the opid:

is the root object id

does not exist in this document

pub fn hydrate(&self, heads: Option<&[ChangeHash]>) -> Value