Trait automerge::sync::SyncDoc

pub trait SyncDoc {
    // Required methods
    fn generate_sync_message(&self, sync_state: &mut State) -> Option<Message>;
    fn receive_sync_message(
        &mut self,
        sync_state: &mut State,
        message: Message
    ) -> Result<(), AutomergeError>;
    fn receive_sync_message_with<Obs: OpObserver>(
        &mut self,
        sync_state: &mut State,
        message: Message,
        op_observer: &mut Obs
    ) -> Result<(), AutomergeError>;
}

A document which can take part in the sync protocol

See the module level documentation for more details.

Required Methods

fn generate_sync_message(&self, sync_state: &mut State) -> Option<Message>

Generate a sync message for the remote peer represented by sync_state

If this returns None then there are no new messages to send, either because we are waiting for an acknolwedgement of an in-flight message, or because the remote is up to date.

fn receive_sync_message( &mut self, sync_state: &mut State, message: Message ) -> Result<(), AutomergeError>

Apply a received sync message to this document and sync_state

fn receive_sync_message_with<Obs: OpObserver>( &mut self, sync_state: &mut State, message: Message, op_observer: &mut Obs ) -> Result<(), AutomergeError>

Apply a received sync message to this document and sync_state, observing any changes with op_observer