Trait automerge::OpObserver

pub trait OpObserver {
    fn insert(
        &mut self,
        doc: &Automerge,
        objid: ExId,
        index: usize,
        tagged_value: (Value<'_>, ExId)
    );
    fn splice_text(
        &mut self,
        _doc: &Automerge,
        _objid: ExId,
        _index: usize,
        _value: &str
    );
    fn put(
        &mut self,
        doc: &Automerge,
        objid: ExId,
        prop: Prop,
        tagged_value: (Value<'_>, ExId),
        conflict: bool
    );
    fn expose(
        &mut self,
        doc: &Automerge,
        objid: ExId,
        prop: Prop,
        tagged_value: (Value<'_>, ExId),
        conflict: bool
    );
    fn increment(
        &mut self,
        doc: &Automerge,
        objid: ExId,
        prop: Prop,
        tagged_value: (i64, ExId)
    );
    fn delete_map(&mut self, doc: &Automerge, objid: ExId, key: &str);
    fn delete_seq(&mut self, doc: &Automerge, objid: ExId, index: usize, num: usize);
    fn branch(&self) -> Self;
    fn merge(&mut self, other: &Self);

    fn flag_conflict(&mut self, _doc: &Automerge, _objid: ExId, _prop: Prop) { ... }
    fn delete(&mut self, doc: &Automerge, objid: ExId, prop: Prop) { ... }
}

An observer of operations applied to the document.

Required Methods

fn insert(
&mut self,
doc: &Automerge,
objid: ExId,
index: usize,
tagged_value: (Value<'_>, ExId)
)

A new value has been inserted into the given object.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been inserted into.

index: the index the new value has been inserted at.

tagged_value: the value that has been inserted and the id of the operation that did the insert.

fn splice_text(
&mut self,
_doc: &Automerge,
_objid: ExId,
_index: usize,
_value: &str
)

fn put(
&mut self,
doc: &Automerge,
objid: ExId,
prop: Prop,
tagged_value: (Value<'_>, ExId),
conflict: bool
)

A new value has been put into the given object.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been put into.

prop: the prop that the value as been put at.

tagged_value: the value that has been put into the object and the id of the operation that did the put.

conflict: whether this put conflicts with other operations.

fn expose(
&mut self,
doc: &Automerge,
objid: ExId,
prop: Prop,
tagged_value: (Value<'_>, ExId),
conflict: bool
)

When a delete op exposes a previously conflicted value Similar to a put op - except for maps, lists and text, edits may already exist and need to be queried

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been put into.

prop: the prop that the value as been put at.

tagged_value: the value that has been put into the object and the id of the operation that did the put.

conflict: whether this put conflicts with other operations.

fn increment(
&mut self,
doc: &Automerge,
objid: ExId,
prop: Prop,
tagged_value: (i64, ExId)
)

A counter has been incremented.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that contains the counter.

prop: they prop that the chounter is at.

tagged_value: the amount the counter has been incremented by, and the the id of the increment operation.

fn delete_map(&mut self, doc: &Automerge, objid: ExId, key: &str)

A map value has beeen deleted.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been deleted in.

key: the map key to be deleted

fn delete_seq(&mut self, doc: &Automerge, objid: ExId, index: usize, num: usize)

A one or more list values have beeen deleted.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been deleted in.

index: the index of the deletion

num: the number of sequential elements deleted

fn branch(&self) -> Self

Branch of a new op_observer later to be merged

Called by AutoCommit when creating a new transaction. Observer branch will be merged on commit() or thrown away on rollback()

fn merge(&mut self, other: &Self)

Merge observed information from a transaction.

Called by AutoCommit on commit()

other: Another Op Observer of the same type

Provided Methods

fn flag_conflict(&mut self, _doc: &Automerge, _objid: ExId, _prop: Prop)

Flag a new conflict on a value without changing it

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been put into.

prop: the prop that the value as been put at.

fn delete(&mut self, doc: &Automerge, objid: ExId, prop: Prop)

A map value has beeen deleted.

doc: a handle to the doc after the op has been inserted, can be used to query information

objid: the object that has been deleted in.

prop: the prop to be deleted