Trait automerge::OpObserver

pub trait OpObserver {
    fn insert(
        &mut self, 
        objid: ExId, 
        index: usize, 
        tagged_value: (Value<'_>, ExId)
    );
    fn put(
        &mut self, 
        objid: ExId, 
        key: Prop, 
        tagged_value: (Value<'_>, ExId), 
        conflict: bool
    );
    fn increment(&mut self, objid: ExId, key: Prop, tagged_value: (i64, ExId));
    fn delete(&mut self, objid: ExId, key: Prop);
}

An observer of operations applied to the document.

Required methods

fn insert(&mut self, objid: ExId, index: usize, tagged_value: (Value<'_>, ExId))

A new value has been inserted into the given object.

objid: the object that has been inserted into.

index: the index the new value has been inserted at.

tagged_value: the value that has been inserted and the id of the operation that did the insert.

fn put(
&mut self,
objid: ExId,
key: Prop,
tagged_value: (Value<'_>, ExId),
conflict: bool
)

A new value has been put into the given object.

objid: the object that has been put into.

key: the key that the value as been put at.

tagged_value: the value that has been put into the object and the id of the operation that did the put.

conflict: whether this put conflicts with other operations.

fn increment(&mut self, objid: ExId, key: Prop, tagged_value: (i64, ExId))

A counter has been incremented.

objid: the object that contains the counter.

key: they key that the chounter is at.

tagged_value: the amount the counter has been incremented by, and the the id of the increment operation.

fn delete(&mut self, objid: ExId, key: Prop)

A value has beeen deleted.

objid: the object that has been deleted in.

key: the key of the value that has been deleted.