Module automerge::transaction

Structs

CommitOptions

Optional metadata for a commit.

Failure

The result of a failed, and rolled back, transaction.

Observed

Success

The result of a successful, and committed, transaction.

Transaction

A transaction on a document. Transactions group operations into a single change so that no other operations can happen in-between.

UnObserved

Traits

Observation

Transactable

A way of mutating a document within a single change.

Type Definitions

Result