automerge::transaction

Struct Transaction

pub struct Transaction<'a> { /* private fields */ }

A transaction on a document. Transactions group operations into a single change so that no other operations can happen in-between.

Created from Automerge::transaction().

Drop

This transaction should be manually committed or rolled back. If not done manually then it will be rolled back when it is dropped. This is to prevent the document being in an unsafe intermediate state. This is consistent with ? error handling.

Implementations

impl<'a> Transaction<'a>

pub fn hash_for_opid(&self, opid: &ExId) -> Option<ChangeHash>

Get the hash of the change that contains the given opid.

Returns none if the opid:

is the root object id

does not exist in this document

is for an operation in this transaction

impl<'a> Transaction<'a>

pub fn get_heads(&self) -> Vec<ChangeHash>

Get the heads of the document before this transaction was started.

pub fn commit(self) -> (Option<ChangeHash>, PatchLog)

Commit the operations performed in this transaction, returning the hashes corresponding to the new heads.

pub fn commit_with( self, options: CommitOptions, ) -> (Option<ChangeHash>, PatchLog)

Commit the operations in this transaction with some options.

let mut doc = Automerge::new();
let mut tx = doc.transaction();
tx.put_object(ROOT, "todos", ObjType::List).unwrap();
let now = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH).unwrap().as_secs() as
i64;
tx.commit_with(CommitOptions::default().with_message("Create todos list").with_time(now));

pub fn rollback(self) -> usize

Undo the operations added in this transaction, returning the number of cancelled operations.