Trait automerge::transaction::Transactable

pub trait Transactable {

    fn pending_ops(&self) -> usize;
    fn put<O: AsRef<ExId>, P: Into<Prop>, V: Into<ScalarValue>>(
        &mut self, 
        obj: O, 
        prop: P, 
        value: V
    ) -> Result<(), AutomergeError>;
    fn put_object<O: AsRef<ExId>, P: Into<Prop>>(
        &mut self, 
        obj: O, 
        prop: P, 
        object: ObjType
    ) -> Result<ExId, AutomergeError>;
    fn insert<O: AsRef<ExId>, V: Into<ScalarValue>>(
        &mut self, 
        obj: O, 
        index: usize, 
        value: V
    ) -> Result<(), AutomergeError>;
    fn insert_object<O: AsRef<ExId>>(
        &mut self, 
        obj: O, 
        index: usize, 
        object: ObjType
    ) -> Result<ExId, AutomergeError>;
    fn increment<O: AsRef<ExId>, P: Into<Prop>>(
        &mut self, 
        obj: O, 
        prop: P, 
        value: i64
    ) -> Result<(), AutomergeError>;
    fn delete<O: AsRef<ExId>, P: Into<Prop>>(
        &mut self, 
        obj: O, 
        prop: P
    ) -> Result<(), AutomergeError>;
    fn splice<O: AsRef<ExId>, V: IntoIterator<Item = ScalarValue>>(
        &mut self, 
        obj: O, 
        pos: usize, 
        del: usize, 
        vals: V
    ) -> Result<(), AutomergeError>;
    fn keys<O: AsRef<ExId>>(&self, obj: O) -> Keys<'_, '_>;
    fn keys_at<O: AsRef<ExId>>(
        &self, 
        obj: O, 
        heads: &[ChangeHash]
    ) -> KeysAt<'_, '_>;
    fn map_range<O: AsRef<ExId>, R: RangeBounds<String>>(
        &self, 
        obj: O, 
        range: R
    ) -> MapRange<'_, R>;
    fn map_range_at<O: AsRef<ExId>, R: RangeBounds<String>>(
        &self, 
        obj: O, 
        range: R, 
        heads: &[ChangeHash]
    ) -> MapRangeAt<'_, R>;
    fn list_range<O: AsRef<ExId>, R: RangeBounds<usize>>(
        &self, 
        obj: O, 
        range: R
    ) -> ListRange<'_, R>;
    fn list_range_at<O: AsRef<ExId>, R: RangeBounds<usize>>(
        &self, 
        obj: O, 
        range: R, 
        heads: &[ChangeHash]
    ) -> ListRangeAt<'_, R>;
    fn values<O: AsRef<ExId>>(&self, obj: O) -> Values<'_>;
    fn values_at<O: AsRef<ExId>>(
        &self, 
        obj: O, 
        heads: &[ChangeHash]
    ) -> Values<'_>;
    fn length<O: AsRef<ExId>>(&self, obj: O) -> usize;
    fn length_at<O: AsRef<ExId>>(&self, obj: O, heads: &[ChangeHash]) -> usize;
    fn object_type<O: AsRef<ExId>>(&self, obj: O) -> Option<ObjType>;
    fn text<O: AsRef<ExId>>(&self, obj: O) -> Result<String, AutomergeError>;
    fn text_at<O: AsRef<ExId>>(
        &self, 
        obj: O, 
        heads: &[ChangeHash]
    ) -> Result<String, AutomergeError>;
    fn get<O: AsRef<ExId>, P: Into<Prop>>(
        &self, 
        obj: O, 
        prop: P
    ) -> Result<Option<(Value<'_>, ExId)>, AutomergeError>;
    fn get_at<O: AsRef<ExId>, P: Into<Prop>>(
        &self, 
        obj: O, 
        prop: P, 
        heads: &[ChangeHash]
    ) -> Result<Option<(Value<'_>, ExId)>, AutomergeError>;
    fn get_all<O: AsRef<ExId>, P: Into<Prop>>(
        &self, 
        obj: O, 
        prop: P
    ) -> Result<Vec<(Value<'_>, ExId)>, AutomergeError>;
    fn get_all_at<O: AsRef<ExId>, P: Into<Prop>>(
        &self, 
        obj: O, 
        prop: P, 
        heads: &[ChangeHash]
    ) -> Result<Vec<(Value<'_>, ExId)>, AutomergeError>;
    fn parent_object<O: AsRef<ExId>>(&self, obj: O) -> Option<(ExId, Prop)>;
    fn parents(&self, obj: ExId) -> Parents<'_>;

    fn splice_text<O: AsRef<ExId>>(
        &mut self, 
        obj: O, 
        pos: usize, 
        del: usize, 
        text: &str
    ) -> Result<(), AutomergeError> { ... }
    fn path_to_object<O: AsRef<ExId>>(&self, obj: O) -> Vec<(ExId, Prop)> { ... }

}

A way of mutating a document within a single change.

Required methods

fn pending_ops(&self) -> usize

Get the number of pending operations in this transaction.

fn put<O: AsRef<ExId>, P: Into<Prop>, V: Into<ScalarValue>>(
&mut self,
obj: O,
prop: P,
value: V
) -> Result<(), AutomergeError>

Set the value of property P to value V in object obj.

Errors

This will return an error if

The object does not exist

The key is the wrong type for the object

The key does not exist in the object

fn put_object<O: AsRef<ExId>, P: Into<Prop>>(
&mut self,
obj: O,
prop: P,
object: ObjType
) -> Result<ExId, AutomergeError>

Set the value of property P to the new object V in object obj.

Returns

The id of the object which was created.

Errors

This will return an error if

The object does not exist

The key is the wrong type for the object

The key does not exist in the object

fn insert<O: AsRef<ExId>, V: Into<ScalarValue>>(
&mut self,
obj: O,
index: usize,
value: V
) -> Result<(), AutomergeError>

Insert a value into a list at the given index.

fn insert_object<O: AsRef<ExId>>(
&mut self,
obj: O,
index: usize,
object: ObjType
) -> Result<ExId, AutomergeError>

Insert an object into a list at the given index.

fn increment<O: AsRef<ExId>, P: Into<Prop>>(
&mut self,
obj: O,
prop: P,
value: i64
) -> Result<(), AutomergeError>

Increment the counter at the prop in the object by value.

fn delete<O: AsRef<ExId>, P: Into<Prop>>(
&mut self,
obj: O,
prop: P
) -> Result<(), AutomergeError>

Delete the value at prop in the object.

fn splice<O: AsRef<ExId>, V: IntoIterator<Item = ScalarValue>>(
&mut self,
obj: O,
pos: usize,
del: usize,
vals: V
) -> Result<(), AutomergeError>

fn keys<O: AsRef<ExId>>(&self, obj: O) -> Keys<'_, '_>

Get the keys of the given object, it should be a map.

fn keys_at<O: AsRef<ExId>>(
&self,
obj: O,
heads: &[ChangeHash]
) -> KeysAt<'_, '_>

Get the keys of the given object at a point in history.

fn map_range<O: AsRef<ExId>, R: RangeBounds<String>>(
&self,
obj: O,
range: R
) -> MapRange<'_, R>

fn map_range_at<O: AsRef<ExId>, R: RangeBounds<String>>(
&self,
obj: O,
range: R,
heads: &[ChangeHash]
) -> MapRangeAt<'_, R>

fn list_range<O: AsRef<ExId>, R: RangeBounds<usize>>(
&self,
obj: O,
range: R
) -> ListRange<'_, R>

fn list_range_at<O: AsRef<ExId>, R: RangeBounds<usize>>(
&self,
obj: O,
range: R,
heads: &[ChangeHash]
) -> ListRangeAt<'_, R>

fn values<O: AsRef<ExId>>(&self, obj: O) -> Values<'_>

fn values_at<O: AsRef<ExId>>(&self, obj: O, heads: &[ChangeHash]) -> Values<'_>

fn length<O: AsRef<ExId>>(&self, obj: O) -> usize

Get the length of the given object.

fn length_at<O: AsRef<ExId>>(&self, obj: O, heads: &[ChangeHash]) -> usize

Get the length of the given object at a point in history.

fn object_type<O: AsRef<ExId>>(&self, obj: O) -> Option<ObjType>

Get type for object

fn text<O: AsRef<ExId>>(&self, obj: O) -> Result<String, AutomergeError>

Get the string that this text object represents.

fn text_at<O: AsRef<ExId>>(
&self,
obj: O,
heads: &[ChangeHash]
) -> Result<String, AutomergeError>

Get the string that this text object represents at a point in history.

fn get<O: AsRef<ExId>, P: Into<Prop>>(
&self,
obj: O,
prop: P
) -> Result<Option<(Value<'_>, ExId)>, AutomergeError>

Get the value at this prop in the object.

fn get_at<O: AsRef<ExId>, P: Into<Prop>>(
&self,
obj: O,
prop: P,
heads: &[ChangeHash]
) -> Result<Option<(Value<'_>, ExId)>, AutomergeError>

Get the value at this prop in the object at a point in history.

fn get_all<O: AsRef<ExId>, P: Into<Prop>>(
&self,
obj: O,
prop: P
) -> Result<Vec<(Value<'_>, ExId)>, AutomergeError>

fn get_all_at<O: AsRef<ExId>, P: Into<Prop>>(
&self,
obj: O,
prop: P,
heads: &[ChangeHash]
) -> Result<Vec<(Value<'_>, ExId)>, AutomergeError>

fn parent_object<O: AsRef<ExId>>(&self, obj: O) -> Option<(ExId, Prop)>

Get the object id of the object that contains this object and the prop that this object is at in that object.

fn parents(&self, obj: ExId) -> Parents<'_>

Provided methods

fn splice_text<O: AsRef<ExId>>(
&mut self,
obj: O,
pos: usize,
del: usize,
text: &str
) -> Result<(), AutomergeError>

Like Self::splice but for text.

fn path_to_object<O: AsRef<ExId>>(&self, obj: O) -> Vec<(ExId, Prop)>